+++ AnalyzerTest/TestSupport.cpp
@@ -62,9 +62,9 @@ std::shared_ptr<std::vector<unsigned char>> TestSupport::GetDataFromTestFilesDir
 
   std::shared_ptr<std::vector<unsigned char>> data(new std::vector<unsigned char>());
 
-  long fileSize = 0;
+  size_t fileSize = 0;
   file.seekg(0, std::ios::end);
-  fileSize = file.tellg();
+  fileSize = static_cast<size_t>(file.tellg());
   file.seekg(0, std::ios::beg);
 
   data->reserve(fileSize);
